refactor: root layout scss 마이그레이션
